Adicion de documentos y Reglas de autorizaciones, hasta grupo 8
queda pendiente las autorizaciones del grupo 9 y determinar la forma de
como se procedera con los topes por año, ej: autoriazaciones de
medicamentos, alto costo y alto costo medicamentos.

var si podemos acomodar el rule engine.

+++ Solution1/Autorizaciones.Domain/Entities/PrestacionAutorizacion.cs
@@ -16,7 +16,7 @@ namespace Sigs.AutorizacionesOnline.Models
         public int PrestacionId { get; set; }
         public virtual Prestacion Prestacion { get; set; }
 
-        public decimal Cantidad { get; set; }
+        public int Cantidad { get; set; }
 
         public decimal Tarifa { get; set; }
